requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
stripe==7.4.0
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1
openpyxl==3.1.5
boto3==1.35.91
openai==1.58.1
pytest==8.3.4
